feat(ledger): deposit endpoint + double-reversal race fix + review hardening
Deposit: POST /accounts/{id}/deposit credits an account from a per-currency
system account (the only account allowed to run negative), so money can enter
the ledger through the API while the system-wide sum stays zero. Idempotent and
currency-checked; reuses the transfer engine.

Fixes (review-driven):
- Double-reversal race (critical): partial UNIQUE index on reverses_tx_id makes
  at most one reversal per transaction race-safe; reverse() now retries on
  transient errors and maps the lost race to the same "already reversed" result.
- Balance trigger enforces SUM(amount_minor)=0 per currency, not just aggregate.
- Overflow-safe post-balances (Math.addExact/subtractExact) + @Max amount cap.
- long pagination offset (deep pages no longer overflow to a 500).
- @Size on Idempotency-Key; ConstraintViolation/HandlerMethodValidation -> 400.
- /admin/reconcile drift query bounded (LIMIT + truncated flag).
- Retry backoff jitter; virtual threads enabled; entry-currency/outbox CHECKs.

Tests: deposit (4), concurrent-reversal race (1), amount/header bounds + deep
pagination (3).

+++ src/main/java/dev/ledger/engine/repository/EntryRepository.java
@@ -45,7 +45,7 @@ public class EntryRepository {
         return sum == null ? 0L : sum;
     }
 
-    public List<Entry> findByAccount(long accountId, int limit, int offset) {
+    public List<Entry> findByAccount(long accountId, int limit, long offset) {
         return jdbc.query(
                 "SELECT " + COLS + " FROM entries WHERE account_id = ? ORDER BY id DESC LIMIT ? OFFSET ?",
                 MAPPER, accountId, limit, offset);
